requirements.txt
<<<<<<< HEAD
torch==2.3.0
torchvision==0.18.0
pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
matplotlib==3.8.4
pandas==2.2.2
scikit-learn==1.4.2
tqdm==4.66.4
streamlit==1.37.1
pytorch-grad-cam==1.4.8
=======
streamlit==1.51.0
torch==2.2.0
torchvision==0.18.0
pillow==12.0.0
numpy==1.26.0
opencv-python-headless==4.9.0.80
matplotlib==3.8.4
pandas==2.3.3
scikit-learn==1.4.2
tqdm==4.66.4
pytorch-grad-cam @ git+https://github.com/jacobgil/pytorch-grad-cam.git
>>>>>>> 4e024fde4f9440d90910c4512c1ec7e2c097cb0e
